+++ tests/test_on_data/test_recording_interfaces.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from platform import python_version
 from sys import platform
+from typing import Literal
 from unittest import skip, skipIf
 
 import numpy as np
@@ -173,6 +174,12 @@ class TestEDFRecordingInterface(RecordingExtractorInterfaceTestMixin, TestCase):
                 # self.check_interface_original_timestamps_inmutability()
 
                 self.check_nwbfile_temporal_alignment()
+
+    # EDF has simultaneous access issues; can't have multiple interfaces open on the same file at once...
+    def check_run_conversion_default_backend_in_nwbconverter(
+        self, nwbfile_path: str, backend: Literal["hdf5", "zarr"] = "hdf5"
+    ):
+        pass
 
 
 class TestIntanRecordingInterface(RecordingExtractorInterfaceTestMixin, TestCase):